requirements.txt
asyncio
cryptography
requests
python-dotenv